requirement.txt
langgraph
langchain
langchain-core
langchain-groq
python-dotenv
pdfplumber